requirements.txt
fastapi
uvicorn
qdrant-client
requests
psycopg2-binary
tqdm
pydantic